requirements.txt
# --- Runtime ---
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0   # load .env in cli.py

# --- Testing  ---
pytest>=8.0,<9.0
responses>=0.25,<1.0   # mock HTTP for requests.Session

